ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.7
pdfplumber==0.10.3
sentence-transformers==2.5.1
scikit-learn==1.4.0
numpy==1.26.3
python-dotenv==1.0.0
reportlab==4.1.0
